refactor: 거래소 지역 판별 로직 공통화
